Started tests for fault injection by testing float to binary conversion
